master: fix some more Visual Studio 2019 errors and warnings.
base_alloc.hpp: add missing def of isnan, fix def of abs (not sure why worked).
base_require.cpp: test abs.

+++ example/general/hes_lu_det.cpp
@@ -57,7 +57,7 @@ bool HesLuDet(void)
     // value of the independent variable
     size_t i;
     for(i = 0; i < n * n; i++)
-        X[i] = Complex(int(i), -int(i) );
+        X[i] = Complex( double(i), -double(i) );
 
     // set the independent variables
     Independent(X);
@@ -70,7 +70,7 @@ bool HesLuDet(void)
     // argument value
     CPPAD_TESTVECTOR(Complex)     x( n * n );
     for(i = 0; i < n * n; i++)
-        x[i] = Complex(2 * i, i);
+        x[i] = Complex( double(2 * i) , double(i) );
 
     // first derivative of the determinant
     CPPAD_TESTVECTOR(Complex) H( n * n * n * n );
